Conflicts:
	burn/tests/integration/sample_models.rb

+++ lib/clipper/mapping/associations/one_to_many.rb
@@ -1,28 +1,70 @@
 module Clipper
-  class Mappings
+  class Mapping
 
     class OneToManyCollection < ::Clipper::Collection
 
-      def initialize(association, parent, children)
+      def initialize(association, parent, children = nil)
+        @loaded = !children.nil?
+        children = [] if children.nil?
+        
         @association = association
         @parent = parent
-        @collection = children
+        @to_enlist = @collection = children
         @mapping = association.associated_mapping
+      end
+
+      def loaded?
+        @loaded
       end
 
       def add(item)
         super
 
-        @parent.__session__.enlist(item) if @parent.__session__
-        @association.set_key(@parent, item)
+        if @parent.__session__
+          @association.set_key(@parent, item)
+          @parent.__session__.enlist(item)
+        else
+          @to_enlist << item
+        end
 
         item
       end
       alias << add
 
+      def each_to_enlist
+        @to_enlist.each { |item| yield item }
+        self
+      end
+
+      def finished_enlisting!
+        @to_enlist = []
+      end
+
+      def each
+        load! unless loaded?
+        @collection.each { |item| yield item }
+      end
+
+      def size
+        load! unless loaded?
+        @collection.size
+      end
+
+      protected
+
+      def load!
+        if @parent.__session__
+          criteria = @association.match_criteria.call(@parent, Clipper::Query::Criteria.new(@mapping))
+
+          @collection = @parent.__session__.find(@mapping, criteria.__options__, criteria.__conditions__) | @to_enlist
+        end
+
+        @loaded = true
+      end
     end
 
     class OneToMany < Association
+      attr_reader :match_criteria
 
       def initialize(mapping, name, mapped_name, &match_criteria)
         raise ArgumentError.new("You must pass a block containing a criteria expression for '#{mapping.name} has_many #{name}'") unless match_criteria
@@ -43,8 +85,7 @@ module Clipper
         criteria = self.match_criteria.call(mapping_criteria, Clipper::Query::Criteria.new(self.associated_mapping))
 
         conditions = criteria.__conditions__
-
-        conditions.field.set(child, conditions.value.field.get(parent))
+        conditions.field.accessor.set(child, conditions.value.field.accessor.get(parent))
       end
 
       def unlink(parent, child)
@@ -52,7 +93,7 @@ module Clipper
         criteria = self.match_criteria.call(mapping_criteria, Clipper::Query::Criteria.new(self.associated_mapping))
 
         conditions = criteria.__conditions__
-        conditions.field.set(child, nil)
+        conditions.field.accessor.set(child, nil)
       end
 
       def load(instance)
@@ -75,35 +116,29 @@ module Clipper
           if data = instance_variable_get(association.instance_variable_name)
             data
           else
-            if __session__
-              instance_variable_set(association.instance_variable_name, OneToManyCollection.new(association, self, association.load(self)))
-            else
-              instance_variable_set(association.instance_variable_name, OneToManyCollection.new(association, self, [])) #Collection.new(association.associated_mapping, [])))
-            end
+            instance_variable_set(association.instance_variable_name, OneToManyCollection.new(association, self))
           end
         end
 
         target.send(:define_method, association.setter) do |new_value|
           raise ArgumentError.new("#{self.class}.#{association.setter} only accepts enumerables") unless new_value.is_a?(Enumerable)
 
-          if items = self.send(association.getter)
-            if __session__
-              items.each do |item|
-                association.unlink(self, item)
-                __session__.enlist(item)
-              end
+          if __session__ and (items = self.send(association.getter))
+            items.each do |item|
+              association.unlink(self, item)
+              __session__.enlist(item)
             end
           end
 
+          collection = OneToManyCollection.new(association, self)
+
           new_value.each do |item|
-            association.set_key(self, item)
-            self.__session__.enlist(item) if self.__session__
+            collection << item
           end
 
-          instance_variable_set(association.instance_variable_name, OneToManyCollection.new(association, self, new_value))
+          instance_variable_set(association.instance_variable_name, collection)
         end
       end
     end
-
   end
 end
